vyčištěno, založen nový projekt a naplněn obsahem z mailu
