+++ .gitignore
@@ -13,3 +13,4 @@ hs_err_pid*
 /Miny02/nbproject/private/
 /Hledanimin/nbproject/private/
 /Hledanimin/build/
+/Miny/nbproject/private/
