--- Hledanimin/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=4418a76f
-build.xml.script.CRC32=27277d20
-build.xml.stylesheet.CRC32=8064a381@1.75.2.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=4418a76f
-nbproject/build-impl.xml.script.CRC32=f00da659
-nbproject/build-impl.xml.stylesheet.CRC32=876e7a8f@1.75.2.48
