+++ Miny/nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=96b32c51
+build.xml.script.CRC32=268172fb
+build.xml.stylesheet.CRC32=8064a381@1.68.1.46
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=96b32c51
+nbproject/build-impl.xml.script.CRC32=db5ad119
+nbproject/build-impl.xml.stylesheet.CRC32=5a01deb7@1.68.1.46
