+++ Miny/src/Tabulka.java
@@ -1,0 +1,13 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+/**
+ *
+ * @author LukeMcNemee
+ */
+public class Tabulka {
+    
+}
